Simulator: tweak steering ranges
Adjust the steering range for both axles. The rear axles is set for
direct angle input with a 1:1 ratio. Maximum range is set to 5 deg,
but actual input signal must be in radians.
The front axles have a steering range corresponding to a full turn
of the steering wheel plus an additional 135 degrees in both
directions (2x 360+90+45, 2x 2.75*pi) and a ratio of 16.5:1 for a
maximum steering angle of 30 degrees.